Move 'compare_feature_classes' function to 'file_utilities.py'
The 'compare_feature_classes' function was removed from its initial location in 'compare_feature_classes.py' and added to 'file_utilities.py' to improve organization. Import paths in the 'removing_overlapping_points.py' file were updated to reflect this change.
